+++ src/main/java/com/user/auth/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.user.auth.controller;
 
-import com.user.auth.dto.request.UserRegisterReqDto;
+import com.user.auth.dto.UserRegisterReqDto;
 import com.user.auth.service.UserService;
 import com.user.auth.util.ResponseMessage;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +8,6 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
