+++ src/main/java/com/user/auth/security/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.user.auth.jwtSecurity;
+package com.user.auth.security;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
