+++ src/main/java/com/user/auth/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.user.auth.jwtSecurity;
+package com.user.auth.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
